+++ src/main/java/com/kurly/cloud/point/api/point/domain/consume/ConsumedPoint.java
@@ -5,13 +5,13 @@ import lombok.Getter;
 @Getter
 public class ConsumedPoint {
   long pointSeq;
-  int consumed;
+  long consumed;
   boolean settle;
 
   /**
    * 기본 생성자.
    */
-  public ConsumedPoint(long pointSeq, int consumed, boolean settle) {
+  public ConsumedPoint(long pointSeq, long consumed, boolean settle) {
     this.pointSeq = pointSeq;
     this.consumed = consumed;
     this.settle = settle;
